What works: !PrintNextAdvert, sh_disable_advert_group.
The situation currently:
The timer for some reason is not running, and you have to manually do
!PrintNextAdvert. Also, the client-side Plugin.Initialise function is
not getting run, so the client can only disable groups with
sh_disable_advert_group "GroupName". Their disabled groups wont even
be saved, so they have to do it every time they join.