Remove laststandsurrender plugin :(
So for some ungodly and demonic reason, some mods heavily conflict
with this plugin in some arcane way, causing this mod to not be loaded at all
on the server, but not the client!

This makes pretty much zero sense, but since the engine is not going to be fixed,
this plugin must perish, since otherwise we get a non-equal number of network
messages on the client and server.